webnn/log.html

<!doctype html>
<meta charset=utf-8>
<title>test element-wise unary operations</title>
<link rel="help" href="https://webmachinelearning.github.io/webnn/#api-mlgraphbuilder-unary">
<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>
<script src="./webnn-polyfill.js"></script>
<div id=log></div>
<script type="module">
  'use strict';
  import * as utils from './utils.js';
  let builder;
  setup(() => {
    const context = navigator.ml.createContext();
    builder = new MLGraphBuilder(context);
  })

  function testUnary(op, input, expected, shape) {
    const x = builder.input('x', {type: 'float32', dimensions: shape});
    const y = builder[op](x);
    const graph = builder.build({y});
    const inputs = {'x': new Float32Array(input)};
    const outputs = {'y': new Float32Array(utils.sizeOfShape(shape))};
    graph.compute(inputs, outputs);
    assert_array_approx_equals_ulp(op, outputs.y, expected, 0);
  }


  test(() => {
    testUnary(
        'log',
        [
          0.07695774575494152,
          0.061607704447576905,
          0.3294265010104085,
          0.2723688623683862,
          0.39012598348926775,
          0.4284013862727071,
          0.025832716797618893,
          0.11563109917646797,
          0.3657843799464965,
          0.05153546344941118,
          0.014544168102282429,
          0.48786174618863865,
          0.2058998699800333,
          0.08353195954105497,
          0.043831308807203206,
          0.272936148346996,
          0.2865260875519381,
          0.24488773299903654,
          0.18569795432436287,
          0.020540962219829373,
          0.32421087855735264,
          0.44436075095447525,
          0.0693654059416835,
          0.2965063657168898,
        ],
        [
          -2.564498764179699, 
          -2.7869683440626147, 
          -1.1104020121055336, 
          -1.3005980197265932, 
          -0.9412855574350177, 
          -0.8476947044830901, 
          -3.656113497474364, 
          -2.157350334906443, 
          -1.0057112450128227, 
          -2.9654850976278078, 
          -4.230565183426074, 
          -0.7177232202584912, 
          -1.5803652964200192, 
          -2.4825259713603685, 
          -3.127406904006987, 
          -1.29851739999228, 
          -1.249925690944984, 
          -1.4069554061150715, 
          -1.6836338267080477, 
          -3.885334229433432, 
          -1.1263611151622355, 
          -0.8111185443013801, 
          -2.6683670092113503, 
          -1.2156865912406256, 
        ],
        [2, 3, 4]);
  }, 'log / 1');
  test(() => {
    testUnary(
        'log',
        [
            2.7048413998894703e+37,
            2.361646895214294e+37,
            2.9919650262927757e+37,
            8.164324192552583e+36,
            3.1257085666545947e+37,
            1.1788312898933793e+37,
            4.080686471910662e+37,
            2.530115073838595e+37,
            4.2154080695896476e+37,
            1.3726953722250843e+36,
            3.3657062753301186e+37,
            1.4610259242863305e+37,
            4.14111321620854e+37,
            3.0185665729527273e+37,
            1.496198921141134e+37,
            3.9036914973580213e+37,
            2.9034528547813314e+37,
            1.0804921407747074e+37,
            2.6210306081590376e+37,
            1.522985615201758e+37,
            3.990683356068259e+37,
            3.0979454196284625e+37,
            2.2974304952338836e+37,
            3.6492014613115686e+37,  
          ],
          [
            86.19069171915568, 
            86.05500765339742, 
            86.29157881175033, 
            84.99283730195862, 
            86.33530943959559, 
            85.36017195616135, 
            86.60191366796211, 
            86.12391322621544, 
            86.6343948412104, 
            83.2098395797617, 
            86.40928627050964, 
            85.57478731759775, 
            86.61661308527572, 
            86.30043051476937, 
            85.59857628016965, 
            86.557571084015, 
            86.2615491091027, 
            85.27306506402503, 
            86.15921604309895, 
            85.61632106960658, 
            86.57961092420194, 
            86.3263875646813, 
            86.02743976322327, 
            86.4901568067159,  
          ],
        [2, 3, 4]);
  }, 'log / 2');
  test(() => {
    testUnary(
        'log',
        [
              0.9885014066932292,
              0.9327760830056802,
              0.8175382924059504,
              0.5881663156075647,
              0.9808643743193626,
              1.0523140285803918,
              1.1929252478851797,
              0.7536850557487607,
              0.6509613800878,
              1.2744082959453285,
              0.6327609724368659,
              0.8111032435695438,
              0.7843204320268429,
              1.101336433836221,
              1.21322952568661,
              0.6170711753064593,
              1.4129398180387835,
              1.7450197688143234,
              0.7421885025970179,
              0.5942823405419944,
              0.8864048443675565,
              0.8937876364342956,
              1.0435079992880505,
              1.2268009585732513,
            ],
            [
              -0.01156521331407722, 
              -0.06959010371710657, 
              -0.20145753644417966, 
              -0.5307455214052083, 
              -0.019321081452242688, 
              0.05099157602651548, 
              0.1764084822131284, 
              -0.2827806961802085, 
              -0.4293049625151788, 
              0.242481989279683, 
              -0.45766253883508423, 
              -0.20935992893769806, 
              -0.242937627812723, 
              0.09652438221325216, 
              0.1932858335676434, 
              -0.4827709046691514, 
              0.34567251117381537, 
              0.5567658844236429, 
              -0.29815202145242786, 
              -0.5204007517708611, 
              -0.12058149778794576, 
              -0.11228707516483474, 
              0.042588113330413145, 
              0.20440993461830897,  
            ],
        [2, 3, 4]);
  }, 'log / 3');  
</script>
